Cleanup: Pull out image code into its own file
